# docs/swagger.yaml
definitions:
  handlers.ProjectInput:
    properties:
      description:
        type: string
      manager_id:
        type: integer
      title:
        type: string
    type: object
...
  models.PriorityEnum:
    enum:
    - low
    - medium
    - high
    type: string
    x-enum-varnames:
    - Low
    - Medium
    - High
  models.Project:
    properties:
      completion_date:
        type: string
      creation_date:
        type: string
      description:
        type: string
      id:
        type: integer
      manager_id:
        type: integer
      title:
        type: string
    type: object
  models.StatusEnum:
    enum:
    - new
    - in_progress
    - done
    type: string
    x-enum-varnames:
    - New
    - InProgress
    - Done
  models.Task:
    properties:
      completion_date:
        type: string
      creation_date:
        type: string
      description:
        type: string
      id:
        type: integer
      priority:
        $ref: '#/definitions/models.PriorityEnum'
      project_id:
        type: integer
      responsible_user_id:
        type: integer
      status:
        $ref: '#/definitions/models.StatusEnum'
      title:
        type: string
    type: object
...
host: localhost:8080
info:
  contact: {}
  description: This is project management service API
  title: Project Management Service API
paths:
  /projects:
    get:
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            items:
              $ref: '#/definitions/models.Project'
            type: array
        "404":
          description: No projects found
          schema:
            type: string
        "500":
          description: Internal server error
          schema:
            type: string
      summary: Get all projects
      tags:
      - projects
    post:
      consumes:
      - application/json
      parameters:
      - description: Project information
        in: body
        name: project
        required: true
        schema:
          $ref: '#/definitions/handlers.ProjectInput'
      produces:
      - application/json
      responses:
        "201":
          description: Project created
          schema:
            type: string
        "400":
          description: Could not decode project
          schema:
            type: string
        "500":
          description: Internal server error
          schema:
            type: string
      summary: Create a project
      tags:
      - projects
  /projects/{id}:
    delete:
      parameters:
      - description: Project ID
        in: path
        name: id
        required: true
        type: integer
      responses:
        "200":
          description: Project deleted
          schema:
            type: string
        "404":
          description: Project not found
          schema:
            type: string
        "500":
          description: Internal server error
          schema:
            type: string
      summary: Delete a project
      tags:
      - projects
    get:
      parameters:
      - description: Project ID
        in: path
        name: id
        required: true
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/models.Project'
        "404":
          description: Project not found
          schema:
            type: string
        "500":
          description: Internal server error
          schema:
            type: string
      summary: Get project by ID
      tags:
      - projects
    put:
      consumes:
      - application/json
      parameters:
      - description: Project ID
        in: path
        name: id
        required: true
        type: integer
      - description: Project information
        in: body
        name: project
        required: true
        schema:
          $ref: '#/definitions/handlers.ProjectInput'
      produces:
      - application/json
      responses:
        "200":
          description: Project updated
          schema:
            type: string
        "400":
          description: Could not decode project
          schema:
            type: string
        "404":
          description: Project not found
          schema:
            type: string
        "500":
          description: Internal server error
          schema:
            type: string
      summary: Update a project
      tags:
      - projects
  /projects/{id}/tasks:
    get:
      parameters:
      - description: Project ID
        in: path
        name: id
        required: true
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            items:
              $ref: '#/definitions/models.Task'
            type: array
        "404":
          description: No tasks found
          schema:
            type: string
        "500":
          description: Internal server error
          schema:
            type: string
      summary: Get all tasks for a project
      tags:
      - projects
  /projects/search:
    get:
      parameters:
      - description: Project title
        in: query
        name: title
        type: string
      - description: Project manager ID
        in: query
        name: manager
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            items:
              $ref: '#/definitions/models.Project'
            type: array
        "400":
          description: Invalid search parameters
          schema:
            type: string
        "404":
          description: No projects found
          schema:
            type: string
        "500":
          description: Internal server error
          schema:
            type: string
      summary: Search projects
      tags:
      - projects
...
swagger: "2.0"